fixed a bug that reset all playcounts. fixed u8 playcount limit.  moved the favorite icon over the gamelist. added ABC and count icons and got rid of the entry in the settings page.  new buttons can be themed but not moved yet.

+++ source/cfg.h
@@ -145,7 +145,7 @@ extern struct CFG CFG;
 extern struct THEME THEME;
 extern u8 ocarinaChoice;
 extern u8 sort;
-extern u8 playcnt;
+extern u16 playcnt;
 extern u8 videoChoice;
 extern u8 languageChoice;
 extern u8 viChoice;
@@ -172,7 +172,7 @@ struct Game_NUM
 {
 	u8 id[8];
 	u8 favorite;
-	u8 count;
+	u16 count;
 };
